Add Additional program options
Now we should be able to use a configuration file.  An example is
located inside of conf/log.conf.

As well, the networking ports/addresses should be configurable.  This
is an important step in making the authority server work again.

To use the conf file, you must use the --conf option on the command
line.